requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
pandas>=2.2,<3.1
scikit-learn>=1.8,<1.9
razorpay>=2.0,<2.1
